requirements.txt
fastapi
uvicorn
python-multipart
boto3
strands-agents
beautifulsoup4
requests
playwright
playwright-stealth
pillow
pillow-heif
PyMuPDF
bedrock-agentcore
